test(streaming): isolate the script retention peak measurement
The tracking allocator is process-wide, so a peak measured in
`streaming_drain.rs` also counts whatever the 28 sibling tests allocate
concurrently — `streaming_memory_is_bounded_not_document_sized` alone allocates
32 MiB. Measured over repeated runs, the same assertion reported 287 bytes when
run alone and 30 KiB to 135 KiB under the default parallel runner, against a
256 KiB budget. That is over half the threshold in scheduling noise, on a
quieter machine than CI.

Moved to its own test binary, which cargo runs as a separate process, so the
measurement is deterministic: 287 bytes with the fix against 6,283,656 without
it. `streaming_scaling.rs` already does this for the same reason.

The remaining peak assertion in `streaming_drain.rs` has a 32 MiB budget and is
not sensitive at this scale.
